split off WrappedReducer to a separate file
